backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6  # For file uploads

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0  # PostgreSQL async driver
psycopg2-binary==2.9.9  # For migrations
alembic==1.13.1  # Migrations (future use)

# OpenAI
openai>=1.59.5

# Environment
python-dotenv>=1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0  # For testing async FastAPI

# Audio pipeline dependencies (reuse from transcription pipeline)
pydub==0.25.1
audioop-lts>=0.2.1
